BDRSPS-925 Add Collections for certain attributes in the survey occurrence v2 mapping
The Attribute and Attribute Value IRIs now have the value in them, instead of the row number.
This greatly reduces the number we create.
A new Collection node joins the Attribute node and the rows that had that attribute.

The following attributes are changed:
* basisOfRecord
* habitat
* preparations
* dataGeneralizations

Same as #272, but for the survey occurrence template