Refactor so agent_iri() is only called when there is actual agent value
Previously, this would be called with agent=None, generate an IRI with "None" in it,
then later in mapping check the field value to see if the IRI should be used.

Now we check the field value first, and only call agent_iri() if there is a value,
and therefore only when the IRI should actually be used in mapping.