BDRHD-27 Use url-quoting for observation value IRIs
This is to prevent IRI collisions when there are different values in the template that slugify to the same string,
e.g. scientificName is "Eucalyptus obliqua (L'Hér)" or "Eucalyptus obliqua L'Hér"
when slugifying these become the Value IRI, with two different rdf:value (raw values),
this is not OK, they need to be different Value IRIs.